fix warning on logs

+++ src/dw_poll.c
@@ -477,7 +477,7 @@ int dw_poll_next(dw_poll_t *p_poll, dw_poll_flags *flags, uint64_t *aux) {
             // dw_poll_wait sets n_cqes
             while (++p_poll->u.iouring_fds.iter < p_poll->u.iouring_fds.n_cqes) {
                 struct io_uring_cqe *cqe = p_poll->u.iouring_fds.cqes[p_poll->u.iouring_fds.iter];
-                dw_log("dw_poll_next: uring cqe res:%lu, flags:%lu, aux:%lu --- ---\n", cqe->res, cqe->flags, io_uring_cqe_get_data64(cqe));
+                dw_log("dw_poll_next: uring cqe res:%d, flags:%u, aux:%lu --- ---\n", cqe->res, cqe->flags, (unsigned long)io_uring_cqe_get_data64(cqe));
 
                 if (cqe->flags & IORING_CQE_F_MORE) {
                     dw_log("URING: MORE set ignoring, the second one will do the job\n");
@@ -501,7 +501,7 @@ int dw_poll_next(dw_poll_t *p_poll, dw_poll_flags *flags, uint64_t *aux) {
                 // re-extract aux data
                 uint64_t cqe_aux = DW_URING_UNPACK_AUX(cqe_userdata);
 
-                dw_log("dw_poll_next: more in detail uring cqe res:%lu, flags:%lu, aux:%lu, cqe_aux:%lu --- ---\n", cqe->res, cqe->flags, io_uring_cqe_get_data64(cqe), cqe_aux);
+                dw_log("dw_poll_next: more in detail uring cqe res:%d, flags:%u, aux:%lu, cqe_aux:%lu --- ---\n", cqe->res, cqe->flags, (unsigned long)io_uring_cqe_get_data64(cqe), (unsigned long)cqe_aux);
                 conn_info_t *conn = cqe_op != DW_URING_OP_POLL ? conn_get_by_id(cqe_aux) : NULL;
                 if (conn && (conn->status == CLOSE || conn->status == NOT_INIT) ){
                     io_uring_cqe_seen(&p_poll->u.iouring_fds.ring, cqe);
